+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and start a new "In Progress" section above it.
 
 ## In progress
 
+- Better argument validation in `resample_spatial`/`resample_cube_spatial` (related to [Open-EO/openeo-python-client#690](https://github.com/Open-EO/openeo-python-client/issues/690))
 
 ## 0.123.0